Updated: main.py
I built a command line interface to allow the user to use the hash table module, which handles the processing and generation of indexes, along with organization. This makes it so that instead of having to go in and hard code functions that fall in line with the rest of the code, the user can just work with the interface provided.